[refactor] Move Custom Exceptions from exception to exception.custom Package
