+++ src/main/java/gp/cnusambe/exception/custom/InvalidPasswordException.java
@@ -1,0 +1,5 @@
+package gp.cnusambe.exception.custom;
+
+public class InvalidPasswordException extends RuntimeException{
+    public InvalidPasswordException() {super("아이디와 비밀번호를 다시 확인해주세요.");}
+}
